claims-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
pydantic
